+++ src/pydng/liblj92/lj92.c
